Langchain/LanGraph-MR/requirements.txt
Flask
flask-cors
python-dotenv
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
